Rename rule to `prefer-type-error`
