+++ rules/prefer-type-error.js
